requirements.txt
droidrun
python-dotenv
streamlit
pillow
google-generativeai
requests
rich